Fetch transactor peer by region
Terraform 0.7.0 introduced a new feature for fetching AMIs via AWS
queries. So, we can leverage this feature to remove the hard-coded AMI
code and query about the right AMI for a given region/virtualization
type.

This commit just update transactor AMI. The same technique could be done
for peers, given they are just barebone ubuntu images, but since peers
are a little more custom-made I prefer to do that on another commit if
necessary.